Refactor for hotel res working
Redefine RPH to be a string so passing in just as sabre gives it such as
'007'
Move Timespan, GuestCounts, Customer struct fields on hotel to functions
that can be appended after the body is composed. Add warnings to those
methods that they may break stuff.

Add note in Translateto add a couple more errors